lint
SCCS-vsn: sys/deprecated/netimp/if_imp.c 4.43
SCCS-vsn: sys/kern/kern_proc.c 4.44
SCCS-vsn: sys/vax/if/if_acc.c 4.26
SCCS-vsn: sys/vax/if/if_ec.c 4.26
SCCS-vsn: sys/kern/uipc_socket2.c 4.30
SCCS-vsn: sys/vm/vm_swap.c 4.12
SCCS-vsn: sys/vax/vax/conf.c 4.61

+++ usr/src/sys/kern/kern_proc.c
@@ -1,4 +1,4 @@
-/*	kern_proc.c	4.43	82/10/21	*/
+/*	kern_proc.c	4.44	82/10/22	*/
 
 #include "../h/param.h"
 #include "../h/systm.h"
@@ -331,7 +331,7 @@ bad:
 	if (bp)
 		brelse(bp);
 	if (bno)
-		rmfree(argmap, ctod(clrnd((int) btoc(NCARGS))), bno);
+		rmfree(argmap, (long)ctod(clrnd((int) btoc(NCARGS))), bno);
 	iput(ip);
 }
